requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
lipsum==0.1.2
requests==2.23.0
tqdm==4.46.0
colorama==0.4.3